# api-gateway/src/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZyvoPay API Gateway",
    "description": "ZyvoPay Payment Gateway API - Complete banking-as-a-service platform for PIX payments, TED transfers, and financial operations.\n\n## Authentication\nAll endpoints (except `/health` and `/api/auth/clients`) require API Key authentication.\n\nProvide your API key via:\n- `X-API-Key` header, or\n- `Authorization: Bearer <api_key>` header\n\n## Rate Limiting\n- Default: 60 requests per minute per API key\n- Rate limit headers are returned in responses\n\n## Webhooks\nConfigure your webhook URL in the client settings to receive real-time event notifications.\n",
    "version": "1.0.0",
    "contact": {
      "name": "ZyvoPay Support",
      "email": "support@zyvopay.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.zyvopay.com",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/auth/clients": {
      "post": {
        "summary": "Create client account",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "document",
                  "email"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "legal_name": {
                    "type": "string"
                  },
                  "document": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "webhook_url": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Client created successfully"
          }
        }
      }
    },
...
    "/api/auth/api-keys": {
      "post": {
        "summary": "Generate new API key",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "scopes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "read",
                      "write"
                    ]
                  },
                  "expires_in_days": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key generated"
          }
        }
      },
      "get": {
        "summary": "List API keys",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "List of API keys"
          }
        }
      }
    },
    "/api/auth/api-keys/{id}": {
      "delete": {
        "summary": "Revoke API key",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key revoked"
          }
        }
      }
    },
...
    "/api/backups": {
      "get": {
        "summary": "List transaction backups",
        "tags": [
          "Backups"
        ],
        "parameters": [
          {
            "name": "backup_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of backups"
          }
        }
      }
    },
    "/api/backups/{id}": {
      "get": {
        "summary": "Get backup details",
        "tags": [
          "Backups"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Backup details"
          }
        }
      }
    },
    "/api/backups/{id}/download": {
      "get": {
        "summary": "Download backup file",
        "tags": [
          "Backups"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Backup download URL"
          }
        }
      }
    },
...
  }
}